% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xinan.R
\name{xinan}
\alias{xinan}
\title{xinan function
This function allow you to calculate the multiplication between two numbers.
The fomula used in the function is simple.}
\usage{
xinan(x, y)
}
\arguments{
\item{x}{enter the number, eg. 2}

\item{y}{enter the number, eg. 5}
}
\description{
xinan function
This function allow you to calculate the multiplication between two numbers.
The fomula used in the function is simple.
}
\examples{
xinan(2,5)

}
\keyword{xinan}
